+++ src/ennoia.cpp
@@ -4,8 +4,9 @@
 #include<cmath>
 #include<SDL3/SDL.h>
 #include<SDL3/SDL_main.h>
+#include"types.h"
 
-#include"debug/macros.cpp"
+#include"debug/errors.cpp"
 #include"vectors/vec2.cpp"
 #include"vectors/vec3.cpp"
 #include"vectors/vec4.cpp"