+++ src/test.cpp
@@ -1,0 +1,14 @@
+
+#include"ennoia.cpp"
+
+
+int main() {
+  Ennoia e("TEST");
+
+  while (!e.ShouldClose()) {
+
+  }
+
+  return 0;
+}
+