+++ src/types.h
@@ -1,0 +1,6 @@
+
+#include <stdbool.h>
+
+typedef std::string String;
+typedef std::string_view StringView;
+